requirements.txt
fastapi
uvicorn
gridspace-guava
python-dotenv